Rename to tralalere_assets
